requirements.txt
torch>=2.0
transformers>=4.35
datasets
trl>=0.7
peft>=0.6
accelerate
numpy
pandas
matplotlib
scikit-learn
pyyaml
tqdm
